Use seccomp_unotify to catch syscalls that violate filterss
The main benefit of seccomp_unotify is an ability to resume the trapped syscall.
We have seen a case when a blocked syscall was in the msan internals that
prevented the Go runtime from reaching the SIGSYS handler and producing a panic
report.

Another benefit of this approach is an ability to run a target application and
find out all syscalls that have to be handled.

PiperOrigin-RevId: 713139953